modify some data schema

+++ trunk/cakephp/app/Controller/LendController.php
@@ -6,7 +6,8 @@ class LendController extends AppController {
 
     public function index(){
 		$lend_records = $this->Lend_Record->find('all');
-    	$this->set('lend_records', $lend_records);  	 
+    	$this->set('lend_records', $lend_records);  
+    	$this->set('lend_status', $this->Lendfunc->lead_status());		
     } 
 
 	public function lend_operation() {
@@ -23,28 +24,32 @@ class LendController extends AppController {
 					if (sizeof($this->data) >1) {
 						foreach (array_keys($this->data) as $key)	{
 							if ($key !== 'Lend_Record') {
-								$lend_books = $this->data[$key];
-								$lend_books['person_id'] = $this->data['Lend_Record']['person_id'];
-								$lend_books['status'] = 'C';
-								$lend_books['lend_date'] = date('Y-m-d H:i:s');
-								$lend_books['create_time'] = $lend_books['lend_date'];
-								$ret = $this->Lend_Record->save($lend_books);
-								if ($ret !== false) {
-									$record_id = $ret["Lend_Record"]["id"];
-									$book_instance_modi = array();
-									$book_instance_modi['id'] = $lend_books['book_instance_id'];
-									$book_instance_modi['book_status'] = 2;
-									$book_instance_modi['s_return_date'] = $lend_books['s_return_date'];
-									$ret = $this->Book_Instance->save($book_instance_modi);
-									if ($ret === false) {
-										$this->Lend_Record->delete($record_id);
+								$book_status = $this->Book_Instance->find('all', array('conditions'=>array('book_status'=>1)));
+								if (!empty($book_status)) {
+									$lend_books = $this->data[$key];
+									$lend_books['person_id'] = $this->data['Lend_Record']['person_id'];
+									$lend_books['status'] = 'C';
+									$lend_books['lend_time'] = date('Y-m-d H:i:s');
+									$lend_books['create_time'] = $lend_books['lend_time'];
+									$ret = $this->Lend_Record->save($lend_books);
+									if ($ret !== false) {
+										$record_id = $ret["Lend_Record"]["id"];
+										$book_instance_modi = array();
+										$book_instance_modi['id'] = $lend_books['book_instance_id'];
+										$book_instance_modi['book_status'] = 2;
+										$book_instance_modi['s_return_date'] = $lend_books['s_return_date'];
+										$book_instance_modi['reserve_person_id'] = null;
+										$ret = $this->Book_Instance->save($book_instance_modi);
+										if ($ret === false) {
+											$this->Lend_Record->delete($record_id);
+										}
 									}
 								}
 							}
 						}
 					}
-					$lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $person_info['Person']['id'], 'return_date' => null, 'status in ("C", "R", "E")') ));
-					$over_lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $person_info['Person']['id'], 'return_date' => null, 'book_instance.s_return_date > current_timestamp', 'status in ("C", "R", "E")') ));
+					$lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $person_info['Person']['id'], 'return_time' => null, 'status in ("C", "R", "E")') ));
+					$over_lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $person_info['Person']['id'], 'return_time' => null, 'book_instance.s_return_date > current_timestamp', 'status in ("C", "R", "E")') ));
 				}
 			}
 		}
@@ -58,13 +63,13 @@ class LendController extends AppController {
 		$over_lend_records = array();
 		$person_info = array();
 		if (!empty($this->data)) {
-			$return_record = $this->Lend_Record->find('all', array('condition' => array('status' => 'C')));
+			$return_record = $this->Lend_Record->find('all', array('conditions' => array('status' => 'C')));
 			if (($return_record !== false) && (!empty($return_record))) {
 				$return_record = $return_record[0];
 				$return_rec = array();
 				$return_rec["id"] = $return_record["Lend_Record"]["id"];
-				$return_rec["status"] = 'R';
-				$return_rec["return_date"] = date('Y-m-d H:i:s');
+				$return_rec["status"] = 'I';
+				$return_rec["return_time"] = date('Y-m-d H:i:s');
 				$ret = $this->Lend_Record->save($return_rec);
 				if ($ret !== false) {
 					$record_id = $ret["Lend_Record"]["id"];
@@ -72,6 +77,13 @@ class LendController extends AppController {
 					$book_instance_modi['id'] = $return_record["Lend_Record"]['book_instance_id'];
 					$book_instance_modi['book_status'] = 1;
 					$book_instance_modi['s_return_date'] = null;
+					$reserve_rec = $this->Lend_Record->find('all', array('conditions' =>array('status' => 'R'), 'order' => 'reserve_time'));
+					var_Dump($reserve_rec);
+					if (!empty($reserve_rec)) {
+						$reserve_rec = $reserve_rec[0];
+						$book_instance_modi['book_status'] = 6;
+						$book_instance_modi['reserve_person_id'] = $reserve_rec["Lend_Record"]['person_id'];
+					}
 					$ret = $this->Book_Instance->save($book_instance_modi);
 					if ($ret === false) {
 						$this->Lend_Record->save($return_record["Lend_Record"]);
@@ -79,8 +91,8 @@ class LendController extends AppController {
 				}
 				$this->Person->id = $return_record['Person']['id'];
 				$person_info = $this->Person->read();
-				$lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $return_record['Person']['id'], 'return_date' => null, 'status in ("C", "R", "E")') ));
-				$over_lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $return_record['Person']['id'], 'return_date' => null, 'book_instance.s_return_date > current_timestamp', 'status in ("C", "R", "E")') ));
+				$lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $return_record['Person']['id'], 'return_time' => null, 'status in ("C", "R", "E")') ));
+				$over_lend_records = $this->Lend_Record->find('all',array('conditions' => array('person_id' => $return_record['Person']['id'], 'return_time' => null, 'book_instance.s_return_date > current_timestamp', 'status in ("C", "R", "E")') ));
 			}
 		}
 		$this->set('person_info', $person_info);
@@ -96,23 +108,6 @@ class LendController extends AppController {
 		$this->Book_Instance->id = $this->data["book_instance_id"];
 		$book = $this->Book_Instance->read();
 		$error_code = $this->Lendfunc->book_auth($person, $book);
-		/*if ($person === false) {
-			$error_code = 1;
-		}
-		else {
-			if ($book === false) {
-				$error_code = 2;
-			}
-			else {
-				if ($book['Book_Instance']['is_lend'] != 'Y') {
-					$error_code = 4;
-				}else if ($book['Book_Instance']['book_status'] != 1) {
-					$error_code = 3;
-				}else if ($book['Book_Instance']['level_id'] > $person['Person']['level_id']) {
-					$error_code = 5;
-				}
-			}
-		}*/
 		if ($error_code != 0) {
 			$person = array();
 			$book = array();
